Refactored conditions
